refactor: restructure into modular architecture
