+++ modules/__init__.py
@@ -1,0 +1,2 @@
+"""High-level package exposing LLM UI components."""
+__all__ = ["parser", "retrieval", "prompting", "logic", "ui"]
