+++ modules/parser/__init__.py
@@ -1,0 +1,2 @@
+from .gothic_rag_ui_parser import md_to_pages, normalize
+__all__ = ["md_to_pages", "normalize"]
