+++ modules/parser/gothic_rag_ui_parser.py
