+++ modules/prompting/__init__.py
@@ -1,0 +1,15 @@
+from .prompt_enhancer import enhance_prompt
+from .prompt_templates import SYSTEM_PROMPT, FEWSHOT_QA
+from .persona_selector import select_persona
+from .style_prompting import apply_style
+from .meta_response import with_introduction
+from .response_formatter import weave_quotes
+__all__ = [
+    "enhance_prompt",
+    "SYSTEM_PROMPT",
+    "FEWSHOT_QA",
+    "select_persona",
+    "apply_style",
+    "with_introduction",
+    "weave_quotes",
+]
