+++ modules/prompting/prompt_templates.py
@@ -1,0 +1,25 @@
+"""Prompt templates and few-shot examples for the assistant."""
+from typing import List, Tuple
+
+# H1.1: system prompt in a more natural human tone
+SYSTEM_PROMPT = (
+    "Cześć! Jestem Pomocnik Gothic LARP — twój serdeczny przewodnik po zasadach "
+    "i świecie gry. Mówię prostym językiem, korzystam tylko z przekazanego "
+    "kontekstu i staram się być pomocny jak towarzysz przy ognisku."
+)
+
+# H1.2: few-shot Q&A examples for context prompting
+FEWSHOT_QA: List[Tuple[str, str]] = [
+    (
+        "Jak wygląda walka w nocy?",
+        "W nocy walka staje się trudniejsza, bo słabsza widoczność wpływa na pewność ciosów.",
+    ),
+    (
+        "Czy mogę stworzyć własną runę?",
+        "Tworzenie nowych run to rzadkość, ale doświadczeni magowie mogą się tego podjąć.",
+    ),
+    (
+        "Ile trwa nauka dwuręcznego miecza?",
+        "Opanowanie walki dwuręcznym mieczem wymaga wielu miesięcy treningu i dobrej kondycji.",
+    ),
+]
