+++ modules/ui/__init__.py
@@ -1,0 +1,3 @@
+from .app import build_app
+from .gradio_ui import build_app as build_debug_app
+__all__ = ["build_app", "build_debug_app"]
