+++ settings.yaml
@@ -1,0 +1,1 @@
+confidence_threshold: 0.3
